Finish text integration
