+++ index.js
@@ -1,19 +1,19 @@
 require('dotenv').config()
 
 const robots = {
-    userInput: require('./robots/user-input.js'),
-    text: require('./robots/text.js')
+    input: require('./robots/input.js'),
+    text: require('./robots/text.js'),
+    state: require('./robots/state.js')
 }
 
 async function start() {
-    const content = {
-        maximumSentences: 7
-    }
+    robots.input()
+    await robots.text()
 
-    robots.userInput(content)
-    await robots.text(content)
-
-    console.log(JSON.stringify(content, null, 2))
+    const content = robots.state.load()
+    console.dir(content, {
+        depth: null
+    })
 }
 
 start()
