+++ robots/input.js
@@ -1,0 +1,35 @@
+const readline = require('readline-sync')
+const state = require('./state.js')
+
+function robot() {
+    const content = {
+        maximumSentences: 7
+    }
+
+    content.searchTerm = askAndReturnSearchTerm()
+    content.prefix = askAndReturnPrefix()
+    content.lang = askAndReturnLang()
+    state.save(content)
+
+    function askAndReturnSearchTerm() {
+        return readline.question('Type a Wikipedia search term: ')
+    }
+
+    function askAndReturnPrefix() {
+        const prefixes = ['Who is', 'What is', 'The history of']
+        const selectedPrefixIndex = readline.keyInSelect(prefixes)
+        const selectedPrefixText = prefixes[selectedPrefixIndex]
+
+        return selectedPrefixText
+    }
+
+    function askAndReturnLang() {
+        const langs = ['en', 'pt']
+        const selectedLangIndex = readline.keyInSelect(langs)
+        const selectedLangText = langs[selectedLangIndex]
+
+        return selectedLangText
+    }
+}
+
+module.exports = robot
